remove trailing slash from http cookie uri-path
The path attribute in the sysauth cookie sent by the Luci login page to
the browser has a trailing slash: "path=/cgi-bin/luci/".  Section 5.1.4
of RFC6265 ("HTTP State Management Mechanism") has this to say about
path formatting:

	4.  Output the characters of the uri-path from the first character
	    up to, but not including, the right-most %x2F ("/").

Remove the trailing slash while formatting the cookie string.

NetSurf is an example of a browser that has rather strict validation
checks and rejects this cookie.

Signed-off-by: Eric Kinzie <ekinzie@labn.net>